mod gan

+++ HW4/utils/models.py
@@ -63,7 +63,7 @@ class CNNGAN(nn.Module):
         self.hidden = args.hidden
 
         self.decoder = nn.Sequential(
-                        nn.Conv2d(self.hidden/9, 8, 3, 1, 1),
+                        nn.Conv2d(int(self.hidden/9), 8, 3, 1, 1),
                         nn.LeakyReLU(),
                         nn.ConvTranspose2d(8, 8, 2, 2, 1),
                         nn.Conv2d(8, 8, 3, 1, 1),
@@ -96,7 +96,7 @@ class CNNGAN(nn.Module):
         #                 nn.Linear(50, 1),
         #             )
     def get_decoding(self, z):
-        return self.decoder(z.view(z.shape[0], self.hidden/9, 3, 3)).squeeze(1)
+        return self.decoder(z.view(z.shape[0], int(self.hidden/9), 3, 3)).squeeze(1)
 
     def get_discrim(self, x):
         return self.discrim(x.view(-1, 1,28,28))
